scripts/script_requirements.txt
click
PyGithub
